# docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API de catálogo musical para o app Flutter. As rotas de autenticação (/api/auth/*) são gerenciadas pelo Better Auth com suporte simultâneo a Bearer Token e Cookie HttpOnly.",
    "title": "cardoso-sound-api",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/favorites": {
      "get": {
        "description": "Recupera a listagem paginada de faixas favoritadas pelo usuário ativo, ordenadas pela data de inclusão (favoritedAt DESC).",
        "operationId": "listFavorites",
        "parameters": [
          {
            "description": "Número da página (>= 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 9007199254740991,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Quantidade de itens por página (1..100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "description": "Lista de faixas favoritadas da página solicitada",
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "album": {
                            "description": "Nome do álbum ou null",
                            "nullable": true,
                            "type": "string"
                          },
                          "artist": {
                            "additionalProperties": false,
                            "description": "Resumo do artista autor da faixa",
                            "properties": {
                              "avatarUrl": {
                                "description": "URL do avatar do artista ou null",
                                "nullable": true,
                                "type": "string"
                              },
                              "id": {
                                "description": "Identificador único do artista (UUID v4)",
                                "format": "uuid",
                                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                                "type": "string"
                              },
                              "name": {
                                "description": "Nome do artista",
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "avatarUrl"
                            ],
                            "type": "object"
                          },
                          "audioUrl": {
                            "description": "URL pública direta de reprodução da faixa (SoundHelix)",
                            "format": "uri",
                            "type": "string"
                          },
                          "coverUrl": {
                            "description": "URL da capa da faixa ou null",
                            "nullable": true,
                            "type": "string"
                          },
                          "createdAt": {
                            "description": "Data de criação da faixa em formato ISO 8601 UTC",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                            "type": "string"
                          },
                          "durationSeconds": {
                            "description": "Duração da faixa em segundos",
                            "exclusiveMinimum": true,
                            "maximum": 9007199254740991,
                            "minimum": 0,
                            "type": "integer"
                          },
                          "favoritedAt": {
                            "description": "Data em que a faixa foi favoritada em formato ISO 8601 UTC",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                            "type": "string"
                          },
                          "genre": {
                            "description": "Gênero musical da faixa",
                            "enum": [
                              "rock",
                              "pop",
                              "electronic",
                              "hip-hop",
                              "jazz",
                              "lo-fi"
                            ],
                            "type": "string"
                          },
                          "id": {
                            "description": "Identificador único da faixa (UUID v4)",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                            "type": "string"
                          },
                          "title": {
                            "description": "Título da faixa musical",
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "album",
                          "genre",
                          "durationSeconds",
                          "coverUrl",
                          "audioUrl",
                          "artist",
                          "createdAt",
                          "favoritedAt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "description": "Metadados de paginação da listagem",
                      "properties": {
                        "hasNext": {
                          "description": "Indica se existe uma próxima página",
                          "type": "boolean"
                        },
                        "hasPrev": {
                          "description": "Indica se existe uma página anterior",
                          "type": "boolean"
                        },
                        "limit": {
                          "description": "Limite de itens solicitados por página",
                          "maximum": 9007199254740991,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "page": {
                          "description": "Página atual da listagem",
                          "maximum": 9007199254740991,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "total": {
                          "description": "Total geral de itens encontrados",
                          "maximum": 9007199254740991,
                          "minimum": 0,
                          "type": "integer"
                        },
                        "totalPages": {
                          "description": "Total de páginas disponíveis",
                          "maximum": 9007199254740991,
                          "minimum": 1,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPages",
                        "hasNext",
                        "hasPrev"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "details": {
                      "description": "Detalhes adicionais ou issues de validação RFC 7807",
                      "nullable": true
                    },
                    "error": {
                      "description": "Identificador canônico do erro",
                      "type": "string"
                    },
                    "message": {
                      "description": "Mensagem descritiva da falha",
                      "type": "string"
                    },
                    "statusCode": {
                      "description": "Código de status HTTP",
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "details"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "Lista as faixas favoritadas pelo usuário autenticado",
        "tags": [
          "Library"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://localhost:3333"
    }
  ],
...
}